Liquibase

+++ src/main/java/com/piseth/java/school/phoneshop/dto/ProductDisplayDTO.java
@@ -1,5 +1,7 @@
 package com.piseth.java.school.phoneshop.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -8,5 +10,5 @@ public class ProductDisplayDTO {
 	private String name;
 	private String model;
 	private String color;
-	private Double salePrice;
+	private BigDecimal salePrice;
 }
